// frontend/src/app/api/model/planItem.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanItemSourceType } from './planItemSourceType';


export interface PlanItem { 
    id: number;
    sourceType: PlanItemSourceType;
    sourceId: number;
    sourceTitle: string;
    done: boolean;
    sourceAvailable: boolean;
}
export namespace PlanItem {
}