// frontend/src/app/api/model/planItemSourceType.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PlanItemSourceType = {
    Task: 'TASK',
    Habit: 'HABIT',
    LearningResource: 'LEARNING_RESOURCE',
} as const;
export type PlanItemSourceType = typeof PlanItemSourceType[keyof typeof PlanItemSourceType];